Update: Jokers & The Shop
Howdy! I went ahead and cooked up a little update JUST FOR YOU!!! 

This update finally adds, like, the one thing Balatro is known for: Jokers! Currently, every Joker in the game is directly taken from Balatro, and they all function exactly as you'd expect (except Baron. I'll get to that one later.) It's pretty rad!

I also went ahead and added a shop!! A shop for you to buy Jokers with using The Money! The Money that YOU earned by WINNING those rounds! (You're such a gamer, you!)
Currently the money system is VERY unsophisticated, but at the very least it exists and functions (I think?)

Lastly, I went ahead and rewrote most of the code in main since, tbh, it was kind of disgusting and hard to read! Now it isn't quite as disgusting, nor is it as hard to read! You're welcome!!!

also i renamed Johnklers to Joker because the 3 extra characters seemed like a pain in the ass okay bye bye